requirements.txt
torch==1.13.1+cu117
torchvision==0.14.1+cu117
addict
future
lmdb
numpy<2.0.0
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
timm 
einops
natsort
joblib
wandb
lpips
matplotlib
seaborn
scikit-learn
basicsr